Move setProject out of withAuth so self-hosted still gets project ref
